storagemanage/requirements.txt
asgiref==3.5.0
boto3==1.21.21
botocore==1.24.21
Django==4.0.3
django-storages==1.12.3
djangorestframework==3.13.1
jmespath==1.0.0
python-dateutil==2.8.2
pytz==2021.3
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
